requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23
pandas>=1.5
matplotlib>=3.6
tqdm>=4.65
scikit-learn>=1.2
Pillow>=9.1

# Classical features
scikit-image>=0.20

# Optional (for configs/report)
pyyaml>=6.0